+++ Coach_Me/flask_app/templates/dashboard_coach.html
@@ -1,13 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
 
-  <head>
+<head>
 
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="description" content="">
     <meta name="author" content="">
-    <link href="https://fonts.googleapis.com/css?family=Poppins:100,100i,200,200i,300,300i,400,400i,500,500i,600,600i,700,700i,800,800i,900,900i&display=swap" rel="stylesheet">
+    <link
+        href="https://fonts.googleapis.com/css?family=Poppins:100,100i,200,200i,300,300i,400,400i,500,500i,600,600i,700,700i,800,800i,900,900i&display=swap"
+        rel="stylesheet">
 
     <title>Dashboard coach</title>
 
@@ -19,24 +21,24 @@
 
     <link rel="stylesheet" href="{{ url_for('static', filename='assets/css/templatemo-training-studio.css') }}">
 
-    </head>
-    
-    <body>
-    
+</head>
+
+<body>
+
     <!-- ***** Preloader Start ***** -->
     <div id="js-preloader" class="js-preloader">
-      <div class="preloader-inner">
-        <span class="dot"></span>
-        <div class="dots">
-          <span></span>
-          <span></span>
-          <span></span>
+        <div class="preloader-inner">
+            <span class="dot"></span>
+            <div class="dots">
+                <span></span>
+                <span></span>
+                <span></span>
+            </div>
         </div>
-      </div>
     </div>
     <!-- ***** Preloader End ***** -->
-    
-    
+
+
     <!-- ***** Header Area Start ***** -->
     <header class="header-area header-sticky">
         <div class="container">
@@ -52,13 +54,14 @@
                             {%if user.is_valid==1 and user.is_banned==0%}
                             <li class="scroll-to-section"><a href="/programs/new">Create program</a></li>
                             {%endif%}
-                            <li class="scroll-to-section"><a href="">Welcome coach <span class="text-danger">{{user.first_name}}</span></a></li>
+                            <li class="scroll-to-section"><a href="">Welcome coach <span
+                                        class="text-danger">{{user.first_name}}</span></a></li>
                             <li class="main-button">
                                 <form action="/logout" method="post">
-                                <input type="submit" value="Logout" class="btn btn-danger">
-                            </form>
+                                    <input type="submit" value="Logout" class="btn btn-danger">
+                                </form>
                             </li>
-                        </ul>        
+                        </ul>
                         <a class='menu-trigger'>
                             <span>Menu</span>
                         </a>
@@ -71,36 +74,40 @@
     <!-- ***** Header Area End ***** -->
 
     <section class="section" id="schedule">
-            <!-- WORK HERE -->
+        <!-- WORK HERE -->
         <div class="container">
             <hr>
             <!-- *** COACH IS BANNED MESSAGE *** -->
             {%if user.is_banned!=0%}
-        
+
             <div class="container mt-4">
                 <div class="card p-4">
                     <h2 class="mb-3">Important Account Information</h2>
-            
+
                     <p>Dear <span class="text-danger">{{user.first_name}}</span></p>
-            
-                    <p>We regret to inform you that your account on [Platform/Service Name] has been temporarily banned due to [reason for the ban].</p>
-            
-                    <p><strong>Reason for the Ban:</strong> [Explain the reason for the ban, including any violations of terms of service or community guidelines.]</p>
-            
+
+                    <p>We regret to inform you that your account on [Platform/Service Name] has been temporarily banned
+                        due to [reason for the ban].</p>
+
+                    <p><strong>Reason for the Ban:</strong> [Explain the reason for the ban, including any violations of
+                        terms of service or community guidelines.]</p>
+
                     <p><strong>Impact of the Ban:</strong></p>
                     <ul>
                         <li>Your account will be temporarily inaccessible.</li>
                         <li>Contact us for more information.</li>
                     </ul>
-            
-                    <p>If you believe this ban was issued in error or if you have any questions, please don't hesitate to contact our support team at [support email/phone number].</p>
-            
-                    <p>Thank you for your understanding and cooperation. We look forward to resolving this matter as smoothly as possible.</p>
-            
+
+                    <p>If you believe this ban was issued in error or if you have any questions, please don't hesitate
+                        to contact our support team at [support email/phone number].</p>
+
+                    <p>Thank you for your understanding and cooperation. We look forward to resolving this matter as
+                        smoothly as possible.</p>
+
                     <p>Best regards,</p>
                 </div>
             </div>
-            
+
             <!-- *** COACH IS VALID >>> List of programs *** -->
             {%elif user.is_valid==1%}
 
@@ -148,22 +155,23 @@
 
 
 
-            <!-- WORK HERE -->
-        
+        <!-- WORK HERE -->
+
     </section>
 
-    
+
     <!-- ***** Footer Start ***** -->
     <footer>
         <div class="container">
             <div class="row">
                 <div class="col-lg-12">
                     <p>Copyright &copy; 2023 COACHINI
-                    
-                    - Designed by <a rel="nofollow" href="" class="tm-text-link" target="_parent">SHARKS.Limited</a></p>
-                    
+
+                        - Designed by <a rel="nofollow" href="" class="tm-text-link" target="_parent">SHARKS.Limited</a>
+                    </p>
+
                     <!-- You shall support us a little via PayPal to info@templatemo.com -->
-                    
+
                 </div>
             </div>
         </div>
@@ -184,12 +192,13 @@
     <script src="{{ url_for('static', filename='assets/js/scrollreveal.min.js') }}"></script>
     <script src="{{ url_for('static', filename='assets/js/waypoints.min.js') }}"></script>
     <script src="{{ url_for('static', filename='assets/js/jquery.counterup.min.js') }}"></script>
-    <script src="{{ url_for('static', filename='assets/js/imgfix.min.js') }}"></script> 
-    <script src="{{ url_for('static', filename='assets/js/mixitup.js') }}"></script> 
+    <script src="{{ url_for('static', filename='assets/js/imgfix.min.js') }}"></script>
+    <script src="{{ url_for('static', filename='assets/js/mixitup.js') }}"></script>
     <script src="{{ url_for('static', filename='assets/js/accordions.js') }}"></script>
-    
+
     <!-- Global Init -->
     <script src="{{ url_for('static', filename='assets/js/custom.js') }}"></script>
 
 </body>
+
 </html>
