+++ Coach_Me/flask_app/templates/dashboard_user.html
@@ -12,7 +12,7 @@
         rel="stylesheet">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/prog_styles.css') }}">
-    <title>Login and Registration</title>
+    <title>Dashboard</title>
 
 
     <!-- Additional CSS Files -->
@@ -54,6 +54,9 @@
                             <!-- <li class="scroll-to-section"><a href="#top" class="active">Home</a></li>
                             <li class="scroll-to-section"><a href="#trainers">Our Coachs</a></li> -->
                             <!-- <li class="scroll-to-section"><a href="">create</a></li>  -->
+                            <li class="scroll-to-section"><a href="">Welcome <span
+                                        class="text-danger">{{user.first_name}}</span></a></li>
+
                             <li class="main-button">
                                 <form action="/logout" method="post" class="dropdown-item">
                                     <input type="submit" value="Logout" class="btn btn-danger">
@@ -72,36 +75,34 @@
     <!-- ***** Header Area End ***** -->
 
     <section class="section" id="schedule">
-        <div class="container">
-            <!-- WORK HERE -->
-            <div class="card-container p-1">
-                {%for program in all_programs%}
-                <div class="card">
-                    <div class="card-header">{{program.name_of_program}}</div>
-                    <!-- <div class="program_for">Program_for: Obese</div> -->
-                    <div class="card-duration">Duration: {{program.duration}} Weeks</div>
 
-                    <div class="card-description">Description: {{program.description_of_program}}</div>
-                    <div class="number_of_enrolled">{{program.number_of_enrolled}}</div>
-                    <!-- <a href="my_program.html">My Programs</a> -->
+        <!-- WORK HERE -->
+        <!-- <div class="card-container"> -->
+        {%for program in all_programs%}
+        <div class="row">
+            <div class="card col-6">
+                <div class="card-header">{{program.name_of_program}}</div>
+                <div class="card-duration">Duration: {{program.duration}} Weeks</div>
 
-                    <!--This button redirect to my_program.html-->
-
-                    <button class="btn btn-lg btn-info"><a href="/users/user_view/{{program.id}}"
-                            class="no-underline text-dark">Choose</a></button>
-
-                    </a>
-                </div>
-                <div class="card-footer">
+                <div class="card-description">Description: {{program.description_of_program}}</div>
+                <div class="number_of_enrolled">{{program.number_of_enrolled}}</div>
+                <div class="card-footer d-flex justify-content-around align-items-center">
                     <div class="created_by">Coach: {{program.coach_name}}</div>
-                    <div class="created_by">{{program.picture}}</div>
+                    <div class="created_by"><img class="rounded-circle w-25"
+                            src="{{ url_for('static', filename=program.coach_picture.replace('flask_app/static/', '')) }}"
+                            class="card-img-top" alt="Exercise Image"></div>
                 </div>
-                {%endfor%}
 
+                <a href="/users/user_view/{{program.id}}" class="btn btn-info no-underline text-dark">Choose</a>
             </div>
-
-            <!-- WORK HERE -->
         </div>
+
+        {%endfor%}
+
+        <!-- </div> -->
+
+        <!-- WORK HERE -->
+
     </section>
