--- Coach_Me/flask_app/templates/index.html
@@ -1,130 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-    <script src="{{ url_for('static', filename='js/main.js') }}"></script>
-    <title>Login and registration</title>
-</head>
-
-<body>
-    <div class="container p-2">
-        <h1>Welcome! let's train!</h1>
-        <h3>Description</h3>
-        <p>Lorem ipsum dolor, sit amet consectetur adipisicing elit. Soluta saepe ipsa porro itaque tempore cupiditate
-            doloribus non veniam ex ratione laboriosam libero nihil laborum atque et quisquam, delectus molestias
-            officia!</p>
-        <div class="d-flex justify-content-between mt-2">
-            <div class="col-5 card shadow p-3 border">
-                <h1 class="text-center text-danger">REGISTER</h1>
-                <form action="/users/create" method="post" enctype="multipart/form-data">
-                    <div class="form-group mb-3">
-
-                        <label for="">I'm here to:</label>
-                        <select id="options" name="role" class="form-control" onchange="user_user_measures()">
-                            <option id="n" value="">----choose your role----</option>
-                            <option id="c" value="c">Coach</option>
-                            <option id="u" value="u">Train</option>
-                            <option id="a" value="a">Admin</option>
-                        </select>
-                        {% for message in get_flashed_messages(category_filter=["role"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-
-
-
-                    <div class="form-group mb-3">
-
-                        <label for="">First Name:</label>
-                        <input type="text" class="form-control" name="first_name">
-                        {% for message in get_flashed_messages(category_filter=["first_name"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-
-                    <div class="form-group mb-3">
-
-                        <label for="">Last Name:</label>
-                        <input type="text" class="form-control" name="last_name">
-                        {% for message in get_flashed_messages(category_filter=["last_name"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-                    <div class="form-group mb-3">
-
-                        <label for="">Email:</label>
-                        <input type="text" class="form-control" name="email">
-                        {% for message in get_flashed_messages(category_filter=["email"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-
-                    <div class="form-group mb-3 hidden" id="dynamicInputContainer">
-
-                        <label for="dynamicInput">Height:(cm)</label>
-                        <input type="number" id="dynamicInput1" name="height" class="form-control">
-                        {% for message in get_flashed_messages(category_filter=["height"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-
-                    <div class="form-group mb-3 hidden" id="dynamicInputContainer">
-
-                        <label for="dynamicInput">Weight:(kg)</label>
-                        <input type="number" id="dynamicInput2" name="weight" class="form-control">
-                        {% for message in get_flashed_messages(category_filter=["weight"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-
-                    <div class="form-group mb-3">
-
-                        <label for="">Password:</label>
-                        <input type="password" class="form-control" name="password">
-                        {% for message in get_flashed_messages(category_filter=["password"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-                    <div class="form-group mb-3">
-                        <label for="">Confirm Password:</label>
-                        <input type="password" class="form-control" name="confirm_password">
-                    </div>
-                    <div class="form-group mb-3" id="hidden_file">
-
-                        <label for="">Upload you picture:</label>
-                        <input type="file" class="form-control" name="file">
-                        {% for message in get_flashed_messages(category_filter=["file"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-                    <input type="submit" value="Register" class="btn btn-primary btn-lg w-100">
-                </form>
-            </div>
-            <div class="col-5 card shadow p-3 border card_height">
-                <h1 class="text-center text-danger">LOGIN</h1>
-                <form action="/login" method="post">
-
-                    <div class="form-group mb-3">
-
-                        <label for="">Email:</label>
-                        <input type="text" class="form-control" name="email">
-                        {% for message in get_flashed_messages(category_filter=["login"]) %}
-                        <span class="text-danger">{{message}}</span>
-                        {% endfor %}
-                    </div>
-                    <div class="form-group mb-3">
-                        <label for="">Password:</label>
-                        <input type="password" class="form-control" name="password">
-                    </div>
-                    <input type="submit" value="Log In" class="btn btn-secondary btn-lg w-100">
-                </form>
-            </div>
-        </div>
-    </div>
-</body>
-
-</html>
